+++ tests/test_app.py
@@ -29,44 +29,50 @@ async def create_flag(
     feature_flag = await service.create_feature_flag(flag_data)
     return jsonable_encoder(feature_flag)
 
-
-@app.get("/api/feature-flags/{id}")
+@app.get("/api/feature-flags/list")
 async def get_flag(
-    id: str, service: FeatureFlagService = Depends(get_feature_flag_service)
+    skip: int, limit: int, service: FeatureFlagService = Depends(get_feature_flag_service)
 ):
-    feature_flag = await service.get_feature_flag(id)
+    feature_flags = await service.list_feature_flags(limit=limit, skip=skip)
+    return jsonable_encoder(feature_flags)
+
+@app.get("/api/feature-flags/{code}")
+async def get_flag(
+    code: str, service: FeatureFlagService = Depends(get_feature_flag_service)
+):
+    feature_flag = await service.get_feature_flag_by_code(code)
     return jsonable_encoder(feature_flag)
 
 
-@app.put("/api/feature-flags/{id}")
+@app.put("/api/feature-flags/{code}")
 async def update_flag(
-    id: str,
+    code: str,
     flag_data: dict,
     service: FeatureFlagService = Depends(get_feature_flag_service),
 ):
-    return await service.update_feature_flag(id, flag_data)
+    return await service.update_feature_flag(code, flag_data)
 
 
-@app.delete("/api/feature-flags/{id}")
+@app.delete("/api/feature-flags/{code}")
 async def delete_flag(
-    id: str, service: FeatureFlagService = Depends(get_feature_flag_service)
+    code: str, service: FeatureFlagService = Depends(get_feature_flag_service)
 ):
-    return await service.delete_feature_flag(id)
+    return await service.delete_feature_flag(code)
 
 
-@app.post("/api/feature-flags/{flag_id}/enable", response_model=FeatureFlag)
+@app.post("/api/feature-flags/{code}/enable", response_model=FeatureFlag)
 async def enable_feature_flag(
-    flag_id: str, service: FeatureFlagService = Depends(get_feature_flag_service)
+    code: str, service: FeatureFlagService = Depends(get_feature_flag_service)
 ):
-    await service.enable_feature_flag(flag_id)
-    flag = await service.get_feature_flag(flag_id)
+    await service.enable_feature_flag(code)
+    flag = await service.get_feature_flag_by_code(code)
     return flag
 
 
-@app.post("/api/feature-flags/{flag_id}/disable", response_model=FeatureFlag)
+@app.post("/api/feature-flags/{code}/disable", response_model=FeatureFlag)
 async def disable_feature_flag(
-    flag_id: str, service: FeatureFlagService = Depends(get_feature_flag_service)
+    code: str, service: FeatureFlagService = Depends(get_feature_flag_service)
 ):
-    await service.disable_feature_flag(flag_id)
-    flag = await service.get_feature_flag(flag_id)
+    await service.disable_feature_flag(code)
+    flag = await service.get_feature_flag_by_code(code)
     return flag