+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.2.0] - 2024-09-10
+## [Unreleased]
+
+### Added
+
+- Updating enable / disable, update, get based on the code instead of id
+- List feature flags based on the offset and skip
+
+## [0.2.0]
 
 ### Added